Replaces static methods with non static
