+++ ConsoleApp1/Pages/ResourcesFiles/ResourceStatusPage.cs
@@ -1,6 +1,7 @@
 ﻿using ConsoleApp1.Constants;
 using ConsoleApp1.Events.Classes;
 using ConsoleApp1.Events.Interfaces;
+using ConsoleApp1.Shared;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -13,7 +14,7 @@ namespace ConsoleApp1.Pages.ResourceFiles
     {
         public static void ResourceStatus(string resourceType)
         {
-            Extension_Methods.ViewStatus(Storage.ResourcesList, resourceType);
+            new ResourceInfo().ViewStatus(Storage.ResourcesList, resourceType);
 
             Console.WriteLine($"Press any key to navigate back to the menu");
             string key = Events.Classes.Resources.storeKey();
