+++ ConsoleApp1/Shared/DisplayInformation.cs
